2023.01.15 알고리즘 공부

+++ java/src/backjoon/backjoon_10093.java
@@ -6,8 +6,8 @@ import java.util.Scanner;
 public class backjoon_10093 {
     public static void main(String[] args) throws IOException {
     	Scanner sc = new Scanner(System.in);
-    	int a = sc.nextInt();
-    	int b = sc.nextInt();
+    	long a = sc.nextLong();
+    	long b = sc.nextLong();
         if(a>b) {
             System.out.println(a-b-1);
         	for(long i = b+1; i<a; i++) System.out.print(i+" ");
